App files moves and App Folder deleted
