--- app/components/layout.js
@@ -1,12 +1,0 @@
-import {Navbar} from './NavBar'
-
-export default function Layout({ children }) {
-  return (
-    <>
-    <div>
-      <Navbar/>
-      <main>{children}</main>
-    </div>
-    </>
-  )
-}
