--- app/pages/_app.js
@@ -1,9 +1,0 @@
-import Layout from '../components/layout'
-import '../styles/globals.css'
-
-function MyApp({ Component, pageProps }) {
-  return(
-  <Component {...pageProps}/>)
-}
-
-export default MyApp
