--- app/pages/about.js
@@ -1,15 +1,0 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import styles from '../styles/Home.module.css'
-import  { NavBar} from "../components/NavBar";
-
-export default function About() {
-  return (
-   <div>
-        <div>
-          <NavBar/>
-          </div>
-   </div>
-   
-  )
-}
