--- app/pages/home.js
@@ -1,9 +1,0 @@
-export default function Home() {
-    return (
-        <div className="flex flex-col">
-        <NavBar />
-        <HomeContent />
-        </div>
-     
-    )
-  }
