--- app/styles/globals.css
@@ -1,198 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-
-.twitterThemeLinks
-{
-  border-color: rgb(29 161 242);
-  background-color: rgb(29 161 242);
-  box-shadow: 0px 3px 10px rgba(0,0,0,.25);
-}
-
-.twitchThemeLinks
-{
-  background-color: rgb(145 70 255);
-  box-shadow: 0px 3px 10px rgba(0,0,0,.25);
-}
-
-.youtubeThemeLinks
-{
-  background-color: rgb(255, 0, 0);
-  box-shadow: 0px 3px 10px rgba(0,0,0,.25);
-}
-
-.tiktokThemeLinks
-{
-  background: rgb(0,242,234);
-  background: linear-gradient(117deg, rgba(0,242,234,1) 0%, rgba(0,7,8,1) 53%, rgba(255,0,80,1) 100%);
-  box-shadow: 0px 3px 10px rgba(0,0,0,.25);
-}
-
-.instagramThemeLinks
-{
-  background: #d6249f;
-  background: radial-gradient(circle at 30% 107%, #fdf497 0%, #fdf497 5%, #fd5949 45%,#d6249f 60%,#285AEB 90%);
-  box-shadow: 0px 3px 10px rgba(0,0,0,.25);
-}
-
-.githubThemeLinks
-{
-  background: rgb(27,31,34);
-  box-shadow: 0px 3px 10px rgba(0,0,0,.25);
-}
-.adminBG
-{
-  background: rgba(66, 71, 92, 1);
-  
-}
-
-.adminContentsBG
-{
-  background: rgb(249, 246, 238);
-}
-.adminBorders
-{
-  border-color: rgb(249, 246, 238);
-}
-.adminNavBG
-{
-  background: rgba(66, 71, 92, 1);
-}
-.adminNavBG:hover
-{
-  background: rgb(249, 246, 238);
-  border-color: rgb(249, 246, 238);
-}
-
-.GeneralBG
-{
-  background: rgb(2,74,102);
-  background: linear-gradient(313deg, rgba(2,74,102,1) 0%, rgba(7,208,255,1) 59%, rgba(0,212,255,1) 100%);
-}
-.AddNew_Suggestion_Btn
-{
-  @apply rounded-[1.1rem]  border-x-[1rem] lgs:border-x-[5rem] border-y-[0.5rem] text-[20px] font-bold;
-  background: rgb(249, 246, 238);
-  border-color: rgb(249, 246, 238);
-
-}
-.HomeSetting
-{
-  @apply flex flex-col md:flex-row md:h-[750px] md:w-screen  translate-y-[60px] md:translate-y-[120px] md:justify-center gap-5 md:gap-5;
-}
-.SlideLeft{
-  @apply fixed bg-slate-400 w-screen h-screen; 
-  animation:  slideleft 0.3s linear;
-  transition: 1s ease-in;
-}
-@keyframes slideleft{
-  0%{
-    opacity: 0;
-    transform: translatex(600px);
-    
-  }
-  100%{
-    opacity: 100%;
-    transform: translatex(0px);
-  }
-  
-}
-
-.SlideRight{
-  @apply fixed bg-slate-400 w-screen h-screen; 
-  animation:  slideRight 0.3s linear;
-  transition: 1s ease-in;
-}
-@keyframes slideRight{
-  0%{
-    opacity: 0;
-    transform: translatex(0px);
-    
-  }
-  100%{
-    opacity: 100%;
-    transform: translatex(600px);
-  }
-  
-}
-
-.SlideUp{
-  @apply fixed bg-slate-400 w-screen h-screen; 
-  animation:  slideup 0.3s linear;
-  transition: 2s ease-in;
-}
-@keyframes slideup{
-  0%{
-    opacity: 0%;
-    transform: translatey(600px);
-    
-  }
-  100%{
-    opacity: 100%;
-    transform: translatey(0px);
-  }
-  
-}
-
-.SlideLeftMenu
-{
-  animation:  slideleftmenu 0.4s linear;
-  transition: 1s ease-in;
-  transition-delay: 1.1s;
-}
-
-@keyframes slideleftmenu{
-  0%{
-    opacity: 0;
-    transform: translatex(600px);
-    
-  }
-  100%{
-    opacity: 100%;
-    transform: translatex(0px);
-  }
-}
-  
-html,
-body {
-  padding: 0;
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-    
-    color: black;
-    width: 100%;
-    height: 100%;
-    overflow: auto;
-    background: linear-gradient(112.1deg, rgb(32, 38, 57) 11.4%, rgb(63, 76, 119) 70.2%);
-
-    
-}
-
-
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-@media (prefers-color-scheme: dark) {
-  html {
-    color-scheme: white;
-  }
-  body {
-    color: black;
-    width: 100%;
-    height: 100%;
-    background: linear-gradient(112.1deg, rgb(32, 38, 57) 11.4%, rgb(63, 76, 119) 70.2%);
-  }
-
- 
-}
-
-
