+++ packages/react-app/package.json
@@ -20,6 +20,7 @@
     "@apollo/react-hooks": "^4.0.0",
     "@arbitrum/sdk": "^2.0.0",
     "@eth-optimism/sdk": "^1.0.0",
+    "@heroicons/react": "^2.0.11",
     "@portis/web3": "^4.0.5",
     "@ramp-network/ramp-instant-sdk": "^2.2.0",
     "@testing-library/jest-dom": "^5.11.4",
