+++ packages/react-app/src/assets/HomePic1.jpg
