+++ packages/react-app/src/views/About.jsx
@@ -1,0 +1,12 @@
+import  { NavBar} from "../components/ui/NavBar";
+
+export default function About() {
+  return (
+   <div>
+        <div>
+          <NavBar/>
+          </div>
+   </div>
+   
+  )
+}
