+++ packages/react-app/src/views/HomeMain.jsx
@@ -1,0 +1,12 @@
+import { NavBar } from "../components/ui/NavBar";
+import { HomeContent } from "../components/ui/HomeContent";
+
+export default function HomeMain() {
+    return (
+        <div className="flex flex-col">
+        <NavBar />
+        <HomeContent />
+        </div>
+     
+    )
+  }
