+++ packages/react-app/src/views/index.js
@@ -2,3 +2,5 @@ export { default as Home } from "./Home";
 export { default as ExampleUI } from "./ExampleUI";
 export { default as Hints } from "./Hints";
 export { default as Subgraph } from "./Subgraph";
+export { default as HomeMain } from "./HomeMain";
+export { default as Admin} from "./Admin";
