+++ packages/react-app/tailwind.config.js
@@ -3,6 +3,16 @@ module.exports = {
   content: ["./src/**/*.{js,jsx}"],
   theme: {
     extend: {},
+    screens: {
+      ssm: '375px',
+      psm: '412px',
+      sm: '640px',
+      md: '768px',
+      lg: '1024px',
+      lgs: '1180px',
+      xl: '1280px',
+      '2xl':'1536px',
+    }
   },
   plugins: [],
 };
